requirements.txt
asgiref==3.3.1
awesomeversion==21.2.3
beautifulsoup4
cryptography==3.3.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
numpy==1.20.2
pandas==1.2.3
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.9.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
whitenoise==2.0.6
django-rest-auth[with_social]
djangorestframework
markdown
django-filter
django-datatables-view
Pillow
pyimgur
